_cleanup_archive/server_backup_20251221_190415/test_2/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
flask>=3.0.0
flask-cors>=4.0.0
markdown>=3.4.0
weasyprint>=60.0
reportlab>=4.0.0

